# feat: add optional caching of params and constants